fix: check for auto-cpufreq

+++ arch/setup/packages/installation/custom.sh
@@ -4,17 +4,21 @@
 
 install_auto_cpufreq() {
   if laptop-detect > /dev/null; then
-    log_pretty_message "Laptop detected, installing auto_cpufreq"
-    path="$git_clone_path/auto-cpufreq"
-    git clone https://github.com/AdnanHodzic/auto-cpufreq.git "$path"
-    cd "$path" && sudo ./auto-cpufreq-installer
-    sudo auto-cpufreq --install
-    sudo systemctl enable --now auto-cpufreq.service
-    sudo systemctl mask power-profiles-daemon.service
-    curl -sL "https://raw.githubusercontent.com/AdnanHodzic/auto-cpufreq/master/auto-cpufreq.conf-example" | sudo tee /etc/auto-cpufreq.conf
-    sudo sed -i '/^\[battery\]/,/^\[/s/^# scaling_min_freq = 800000/scaling_min_freq = 1600000/' /etc/auto-cpufreq.conf
-    sudo systemctl restart auto-cpufreq.service
-    rm -rf "$path"
+    if ! command -v auto-cpufreq >/dev/null 2>&1; then
+      log_pretty_message "Laptop detected, installing auto_cpufreq"
+      path="$git_clone_path/auto-cpufreq"
+      git clone https://github.com/AdnanHodzic/auto-cpufreq.git "$path"
+      cd "$path" && sudo ./auto-cpufreq-installer
+      sudo auto-cpufreq --install
+      sudo systemctl enable --now auto-cpufreq.service
+      sudo systemctl mask power-profiles-daemon.service
+      curl -sL "https://raw.githubusercontent.com/AdnanHodzic/auto-cpufreq/master/auto-cpufreq.conf-example" | sudo tee /etc/auto-cpufreq.conf
+      sudo sed -i '/^\[battery\]/,/^\[/s/^# scaling_min_freq = 800000/scaling_min_freq = 1600000/' /etc/auto-cpufreq.conf
+      sudo systemctl restart auto-cpufreq.service
+      rm -rf "$path"
+    else
+      log_pretty_message "auto-cpufreq is already installed" ℹ️
+    fi
   else
     log_pretty_message "No laptop detected, skipping auto_cpufreq installation"
   fi
